REST Spring services for quotes and categories

+++ trunk/GoJava7/trentovskyjvova/Kickstarter/src/main/java/ua/com/goit/gojava7/kickstarter/controller/CategoriesController.java
@@ -57,7 +57,7 @@ public class CategoriesController {
 	}
 
 	@RequestMapping("/projects")
-	public ModelAndView projects(@RequestParam(name = "id") int categoryId) {
+	public ModelAndView projects(@RequestParam(name = "id") Long categoryId) {
 		ModelAndView modelAndView = new ModelAndView("projects");
 
 		modelAndView.addObject("categoryId", categoryId);
@@ -65,7 +65,7 @@ public class CategoriesController {
 
 		List<Project> projects = projectDao.getProjects(categoryId);
 		for (Project project : projects) {
-			int amountPledge = 0;
+			Long amountPledge = 0L;
 			for (Payment payment : project.getPayments()) {
 				amountPledge += payment.getPledge();
 			}
@@ -78,7 +78,7 @@ public class CategoriesController {
 	}
 
 	@RequestMapping("/project")
-	public ModelAndView project(@RequestParam(name = "projectId") int projectId) {
+	public ModelAndView project(@RequestParam(name = "projectId") Long projectId) {
 		ModelAndView modelAndView = new ModelAndView("project");
 
 		log.debug("projectId: {}", projectId);
@@ -95,7 +95,7 @@ public class CategoriesController {
 	}
 
 	@RequestMapping("/payments")
-	public ModelAndView payments(@RequestParam(name = "projectId") int projectId) {
+	public ModelAndView payments(@RequestParam(name = "projectId") Long projectId) {
 		ModelAndView modelAndView = new ModelAndView("payments");
 
 		modelAndView.addObject("projectId", projectId);
@@ -109,7 +109,7 @@ public class CategoriesController {
 	}
 
 	@RequestMapping(value = "/question", method = RequestMethod.GET)
-	public ModelAndView questionGet(@RequestParam(name = "projectId") int projectId) {
+	public ModelAndView questionGet(@RequestParam(name = "projectId") Long projectId) {
 		ModelAndView modelAndView = new ModelAndView("question");
 
 		modelAndView.addObject("projectId", projectId);
@@ -119,7 +119,7 @@ public class CategoriesController {
 	}
 	
 	@RequestMapping(value = "/question", method = RequestMethod.POST)
-	public ModelAndView questionPost(@RequestParam(name = "projectId") int projectId,
+	public ModelAndView questionPost(@RequestParam(name = "projectId") Long projectId,
 			@RequestParam(name = "questionText") String questionText) {
 		ModelAndView modelAndView = new ModelAndView("redirect:project.html");
 
@@ -137,9 +137,9 @@ public class CategoriesController {
 	}
 	
 	@RequestMapping(value = "/pledge", method = RequestMethod.GET)
-	public ModelAndView pledgeGet(@RequestParam(name = "projectId") int projectId,
-			@RequestParam(name = "rewardId") int rewardId,
-			@RequestParam(name = "amount") int amount) {
+	public ModelAndView pledgeGet(@RequestParam(name = "projectId") Long projectId,
+			@RequestParam(name = "rewardId") Long rewardId,
+			@RequestParam(name = "amount") Long amount) {
 		ModelAndView modelAndView = new ModelAndView("pledge");
 
 		modelAndView.addObject("projectId", projectId);
@@ -159,9 +159,9 @@ public class CategoriesController {
 	}
 	
 	@RequestMapping(value = "/pledge", method = RequestMethod.POST)
-	public ModelAndView pledgePost(@RequestParam(name = "projectId") int projectId,
-			@RequestParam(name = "rewardId") int rewardId,
-			@RequestParam(name = "amount") int amount,
+	public ModelAndView pledgePost(@RequestParam(name = "projectId") Long projectId,
+			@RequestParam(name = "rewardId") Long rewardId,
+			@RequestParam(name = "amount") Long amount,
 			@RequestParam(name = "name") String name,
 			@RequestParam(name = "cardNumber") String cardNumber) {
 		ModelAndView modelAndView = new ModelAndView("redirect:project.html");
