+++ trunk/GoJava7/trentovskyjvova/Kickstarter/src/main/java/ua/com/goit/gojava7/kickstarter/dao/CategoryDao.java
@@ -7,6 +7,6 @@ import ua.com.goit.gojava7.kickstarter.domain.Category;
 public interface CategoryDao extends Dao {
 	List<Category> getCategories();
 	
-	Category getCategory(int id);
+	Category getCategory(Long id);
 	
 }
